got volume working

+++ Effect.h
@@ -1,9 +1,10 @@
 #ifndef EFFECT_H
 #define EFFECT_H
+#include <cstdint>
 
 class Effect{
     public:
-        virtual float process(float input) = 0;
+        virtual float process(int16_t input) = 0;
 
         virtual ~Effect();
     private:
